start fixing getDetailsPlayerIndependentByGameResponse api

+++ LastGame/src/main/java/com/IGallinari/LastGame/payload/response/DetailsPlayerIndependByGame/ViewGamesDetailsPlayerIndependentByGame.java
@@ -9,5 +9,5 @@ public class ViewGamesDetailsPlayerIndependentByGame {
     private Integer gamesPlayed;
     private Integer jerseyNumber;
     private Float averageScore;
-    private Float seasonScore;
+    private Integer seasonScore;
 }
